feat: add editable header to create entity page
