+++ apps/web/pages/space/[id]/[entityId].tsx
@@ -7,7 +7,7 @@ import { useAccessControl } from '~/modules/auth/use-access-control';
 import { EditableEntityPage } from '~/modules/components/entity/editable-entity-page';
 import { ReadableEntityPage } from '~/modules/components/entity/readable-entity-page';
 import { ReferencedByEntity } from '~/modules/components/entity/types';
-import { Entity, EntityStoreProvider, useEntityStore } from '~/modules/entity';
+import { Entity, EntityStoreProvider } from '~/modules/entity';
 import { Params } from '~/modules/params';
 import { NetworkData } from '~/modules/io';
 import { StorageClient } from '~/modules/services/storage';
@@ -17,16 +17,10 @@ import { NavUtils } from '~/modules/utils';
 import { DEFAULT_PAGE_SIZE } from '~/modules/triple';
 import { Value } from '~/modules/value';
 import { TypesStoreProvider } from '~/modules/type/types-store';
-import { Truncate } from '~/modules/design-system/truncate';
-import { Text } from '~/modules/design-system/text';
-import { Spacer } from '~/modules/design-system/spacer';
-import { EntityPageMetadataHeader } from '~/modules/components/entity-page/entity-page-metadata-header';
-import { Editor } from '~/modules/components/editor/editor';
 import { EntityPageCover } from '~/modules/components/entity/entity-page-cover';
 import { EntityPageContentContainer } from '~/modules/components/entity/entity-page-content-container';
-import { PageStringField } from '~/modules/components/entity/editable-fields';
-import { useActionsStore } from '~/modules/action';
-import { useEditEvents } from '~/modules/components/entity/edit-events';
+import { EditableHeading } from '~/modules/components/entity/editable-entity-header';
+import { fetchForeignTypeTriples, fetchSpaceTypeTriples } from '~/modules/spaces/fetch-types';
 
 interface Props {
   triples: Triple[];
@@ -89,83 +83,6 @@ export default function EntityPage(props: Props) {
         </EntityStoreProvider>
       </TypesStoreProvider>
     </>
-  );
-}
-
-function EditableHeading({
-  spaceId,
-  entityId,
-  name: serverName,
-  triples: serverTriples,
-}: {
-  spaceId: string;
-  entityId: string;
-  name: string;
-  triples: Triple[];
-}) {
-  const { triples: localTriples, update, create, remove } = useEntityStore();
-  const { editable } = useEditable();
-  const { isEditor } = useAccessControl(spaceId);
-  const { actionsFromSpace } = useActionsStore(spaceId);
-
-  const triples = localTriples.length === 0 && actionsFromSpace.length === 0 ? serverTriples : localTriples;
-
-  const nameTriple = Entity.nameTriple(triples);
-  const name = Entity.name(triples) ?? serverName;
-  const types = Entity.types(triples) ?? [];
-
-  const isEditing = editable && isEditor;
-
-  const send = useEditEvents({
-    context: {
-      entityId,
-      spaceId,
-      entityName: name,
-    },
-    api: {
-      create,
-      update,
-      remove,
-    },
-  });
-
-  const onNameChange = (e: React.ChangeEvent<HTMLTextAreaElement>) => {
-    send({
-      type: 'EDIT_ENTITY_NAME',
-      payload: {
-        name: e.target.value,
-        triple: nameTriple,
-      },
-    });
-  };
-
-  return (
-    <div>
-      {isEditing ? (
-        <div>
-          <PageStringField variant="mainPage" placeholder="Entity name..." value={name} onChange={onNameChange} />
-          {/* 
-            This height differs from the readable page height due to how we're using an expandable textarea for editing
-            the entity name. We can't perfectly match the height of the normal <Text /> field with the textarea, so we
-            have to manually adjust the spacing here to remove the layout shift.
-          */}
-          <Spacer height={5.5} />
-        </div>
-      ) : (
-        <div>
-          <Truncate maxLines={3} shouldTruncate>
-            <Text as="h1" variant="mainPage">
-              {name}
-            </Text>
-          </Truncate>
-          <Spacer height={12} />
-        </div>
-      )}
-
-      <EntityPageMetadataHeader id={entityId} spaceId={spaceId} types={types} />
-      <Spacer height={40} />
-      <Editor editable={isEditing} />
-    </div>
   );
 }
 
@@ -254,60 +171,4 @@ export const getServerSideProps: GetServerSideProps<Props> = async context => {
       spaceTypes: [...spaceTypes, ...foreignSpaceTypes],
     },
   };
-};
-
-export const fetchSpaceTypeTriples = async (network: NetworkData.INetwork, spaceId: string) => {
-  /* Fetch all entities with a type of type (e.g. Person / Place / Claim) */
-
-  const { triples } = await network.fetchTriples({
-    query: '',
-    space: spaceId,
-    skip: 0,
-    first: DEFAULT_PAGE_SIZE,
-    filter: [
-      { field: 'attribute-id', value: SYSTEM_IDS.TYPES },
-      {
-        field: 'linked-to',
-        value: SYSTEM_IDS.SCHEMA_TYPE,
-      },
-    ],
-  });
-
-  return triples;
-};
-
-export const fetchForeignTypeTriples = async (network: NetworkData.INetwork, space: Space) => {
-  if (!space.spaceConfigEntityId) {
-    return [];
-  }
-
-  const foreignTypesFromSpaceConfig = await network.fetchTriples({
-    query: '',
-    space: space.id,
-    skip: 0,
-    first: DEFAULT_PAGE_SIZE,
-    filter: [
-      { field: 'entity-id', value: space.spaceConfigEntityId },
-      { field: 'attribute-id', value: SYSTEM_IDS.FOREIGN_TYPES },
-    ],
-  });
-
-  const foreignTypesIds = foreignTypesFromSpaceConfig.triples.map(triple => triple.value.id);
-
-  const foreignTypes = await Promise.all(
-    foreignTypesIds.map(entityId =>
-      network.fetchTriples({
-        query: '',
-        skip: 0,
-        first: DEFAULT_PAGE_SIZE,
-        filter: [
-          { field: 'entity-id', value: entityId },
-          { field: 'attribute-id', value: SYSTEM_IDS.TYPES },
-          { field: 'linked-to', value: SYSTEM_IDS.SCHEMA_TYPE },
-        ],
-      })
-    )
-  );
-
-  return foreignTypes.flatMap(foreignType => foreignType.triples);
 };
